Update main.rb
WOW there's a lot going on there. I would consider moving some of the helper methods to another file. That will clean up the code a little. Remember to add comment to explain what's going on. It's probably a good idea to remove the dead code, your commit history will tell the story.